rename and refocus project
