+++ .github/workflows/merge.yml
@@ -1,4 +1,4 @@
-name: Elevatorian SDLC - Production
+name: Ukol SDLC - Production
 
 on:
   push:
