+++ .github/workflows/pr.yml
@@ -1,4 +1,4 @@
-name: Elevatorian SDLC - Pull Request
+name: Ukol SDLC - Pull Request
 
 on:
   push:
