+++ apps/api/db/knexfile.js
@@ -1,6 +1,6 @@
 module.exports = {
   client: 'pg',
-  connection: 'postgresql://postgres:postgres@db:5432/elevatorian',
+  connection: 'postgresql://postgres:postgres@db:5432/ukol',
   migrations: {
     directory: './migrations',
     tableName: 'knex_migrations',
@@ -10,4 +10,4 @@ module.exports = {
   },
   debug: true,
   acquireConnectionTimeout: 5000,
-}
+};
