+++ apps/api/db/run_migrations.sh
