+++ apps/api/src/app/db/init.ts
@@ -1,15 +1,15 @@
-import Knex from 'knex'
+import Knex from 'knex';
 
-import { Model } from 'objection'
+import { Model } from 'objection';
 
 export const initDb = async (env: any) => {
   const knex = Knex({
     client: 'pg',
-    connection: env.db
+    connection: env.db,
   });
 
   try {
-    console.log(`connecting to db: ${env.db}`)
+    console.log(`connecting to db: ${env.db}`);
     await knex.raw(`select now()`);
     console.log(`db connection successful!`);
   } catch (e) {
