+++ apps/api/src/app/db/models.ts
@@ -1,0 +1,8 @@
+import { Model } from 'objection';
+
+export class Person extends Model {
+  uuid!: string;
+  name!: string;
+
+  static tableName = 'person';
+}
