+++ apps/api/src/app/graphql/resolvers/index.ts
@@ -1,0 +1,9 @@
+import personQueries from './person';
+
+const resolvers = {
+  Query: {
+    ...personQueries,
+  },
+};
+
+export default resolvers;
