+++ apps/api/src/app/graphql/resolvers/person.ts
@@ -1,0 +1,19 @@
+import { Person as PersonModel } from '../../db/models';
+import { Person as PersonGql } from '@ukol/graphql';
+
+const personQueries = {
+  person: async (parent, args, context, info) => {
+    // fetch
+    const model = await PersonModel.query().findOne('uuid', args.uuid);
+
+    // transpose db model to graphql model
+    const gql: PersonGql = {
+      uuid: model.uuid,
+      name: model.name,
+    };
+
+    return gql;
+  },
+};
+
+export default personQueries;
