--- apps/express-app/src/app/db/models.ts
@@ -1,8 +1,0 @@
-import { Model } from 'objection'
-
-export class Person extends Model {
-  uuid!: string
-  name!: string
-
-  static tableName = 'person'
-}
