--- apps/express-app/src/app/graphql/resolvers/index.ts
@@ -1,9 +1,0 @@
-import personQueries from "./person";
-
-const resolvers = {
-  Query: {
-    ...personQueries
-  },
-};
-
-export default resolvers;
