--- apps/web-e2e/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "sourceMap": false,
-    "outDir": "../../dist/out-tsc",
-    "allowJs": true,
-    "types": ["cypress", "node"]
-  },
-  "include": ["src/**/*.ts", "src/**/*.js", "cypress.config.ts"]
-}
