+++ docker-compose.yml
@@ -23,7 +23,7 @@ services:
       - 3333:3333
     environment:
       - PORT=3333
-      - DB_CONNECTION_STRING=postgresql://postgres:postgres@db:5432/elevatorian
+      - DB_CONNECTION_STRING=postgresql://postgres:postgres@db:5432/ukol
     depends_on:
       - db
       - migrate
@@ -34,7 +34,7 @@ services:
       service: nx-base
     command: ash ./apps/express-app/db/run_migrations.sh
     environment:
-      - DB_CONNECTION_STRING=postgresql://postgres:postgres@db:5432/elevatorian
+      - DB_CONNECTION_STRING=postgresql://postgres:postgres@db:5432/ukol
     depends_on:
       - db
 
@@ -43,7 +43,7 @@ services:
     environment:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
-      - POSTGRES_DB=elevatorian
+      - POSTGRES_DB=ukol
     ports:
       - "5432:5432"
     volumes:
