+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "elevatorian",
+  "name": "ukol",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
