+++ workspace.json
@@ -2,9 +2,8 @@
   "$schema": "./node_modules/nx/schemas/workspace-schema.json",
   "version": 2,
   "projects": {
-    "express-app": "apps/express-app",
+    "api": "apps/api",
     "graphql": "libs/graphql",
-    "web": "apps/web",
-    "web-e2e": "apps/web-e2e"
+    "web": "apps/web"
   }
 }
